+++ NWNXLib/Utils.hpp
@@ -52,15 +52,41 @@ CNWSScriptVarTable *GetScriptVarTable(CGameObject *pObject);
 void DestroyGameEffect(CGameEffect* pEffect);
 
 std::string ExtractLocString(CExoLocString& locStr, int32_t nID = 0, uint8_t bGender = 0);
+CExoLocString CreateLocString(const std::string& str, int32_t nID = 0, uint8_t bGender = 0);
 
 template <typename T>
-T PeekMessage(CNWSMessage *pMessage, int32_t offset)
+inline T PeekMessage(CNWSMessage *pMessage, int32_t offset)
 {
     static_assert(std::is_pod<T>::value);
     T value;
     uint8_t *ptr = pMessage->m_pnReadBuffer + pMessage->m_nReadBufferPtr + offset;
     std::memcpy(&value, ptr, sizeof(T));
     return value;
+}
+
+template <>
+inline std::string PeekMessage<std::string>(CNWSMessage *pMessage, int32_t offset)
+{
+    std::string string;
+    auto length = PeekMessage<int32_t>(pMessage, offset);
+
+    string.reserve(length + 1);
+    string.assign(reinterpret_cast<const char*>(pMessage->m_pnReadBuffer + pMessage->m_nReadBufferPtr + offset + 4), length);
+    string[length] = '\0';
+
+    return string;
+}
+
+template <>
+inline CResRef PeekMessage<CResRef>(CNWSMessage *pMessage, int32_t offset)
+{
+    std::string string;
+
+    string.reserve(16 + 1);
+    string.assign(reinterpret_cast<const char*>(pMessage->m_pnReadBuffer + pMessage->m_nReadBufferPtr + offset), 16);
+    string[16] = '\0';
+
+    return CResRef(string);
 }
 
 void AddStealthEvent(int32_t which, ObjectID oidSelf, ObjectID oidTarget);
@@ -76,4 +102,5 @@ int PopScriptContext();
 
 void SetOrientation(CNWSObject *pObject, float facing);
 
+bool IsValidCustomResourceDirectoryAlias(const std::string& filename);
 }