+++ Plugins/Creature/Creature.hpp
@@ -118,14 +118,29 @@ private:
     ArgumentStack SetCriticalRangeOverride      (ArgumentStack&& args);
     ArgumentStack GetCriticalRangeOverride      (ArgumentStack&& args);
     ArgumentStack AddAssociate                  (ArgumentStack&& args);
+    ArgumentStack SetLastItemCasterLevel        (ArgumentStack&& args);
+    ArgumentStack GetLastItemCasterLevel        (ArgumentStack&& args);
+    ArgumentStack GetArmorClassVersus           (ArgumentStack&& args);
+    ArgumentStack SetEffectIconFlashing         (ArgumentStack&& args);
+    ArgumentStack OverrideDamageLevel           (ArgumentStack&& args);
+    ArgumentStack SetEncounter                  (ArgumentStack&& args);
+    ArgumentStack GetEncounter                  (ArgumentStack&& args);
+    ArgumentStack GetIsBartering                (ArgumentStack&& args);
+    ArgumentStack GetWalkAnimation              (ArgumentStack&& args);
+    ArgumentStack SetWalkAnimation              (ArgumentStack&& args);
+    ArgumentStack SetAttackRollOverride         (ArgumentStack&& args);
+    ArgumentStack SetParryAllAttacks            (ArgumentStack&& args);
 
     CNWSCreature *creature(ArgumentStack& args);
+    std::unordered_map<uint8_t, std::unordered_map<ObjectID, int16_t>> m_RollModifier;
+    std::unordered_map<ObjectID, bool> m_ParryAllAttacks;
 
 private:
     static bool s_bAdjustCasterLevel;
     static bool s_bCasterLevelHooksInitialized;
     static bool s_bCriticalMultiplierHooksInitialized;
     static bool s_bCriticalRangeHooksInitialized;
+    static bool s_bResolveAttackRollHookInitialized;
 
     static void InitCasterLevelHooks();
     static void CNWSCreatureStats__GetClassLevel(bool before, CNWSCreatureStats* thisPtr, uint8_t nMultiClass, BOOL bUseNegativeLevel);
@@ -134,6 +149,8 @@ private:
     static void CGameEffect__SetCreator(bool before, CGameEffect*, OBJECT_ID);
     static void InitCriticalMultiplierHook();
     static void InitCriticalRangeHook();
+    static void DoResolveAttackHook(CNWSCreature* thisPtr, CNWSObject* pTarget);
+    static bool InitResolveAttackRollHook();
 
 };