+++ Plugins/Player/Player.hpp
@@ -4,6 +4,8 @@
 #include "Services/Events/Events.hpp"
 #include "Services/Hooks/Hooks.hpp"
 #include "API/ObjectVisualTransformData.hpp"
+#include "API/CExoLocString.hpp"
+#include "Utils.hpp"
 #include <map>
 #include <set>
 
@@ -54,6 +56,13 @@ private:
     ArgumentStack SetCreatureNameOverride           (ArgumentStack&& args);
     ArgumentStack FloatingTextStringOnCreature      (ArgumentStack&& args);
     ArgumentStack ToggleDM                          (ArgumentStack&& args);
+    ArgumentStack SetObjectMouseCursorOverride      (ArgumentStack&& args);
+    ArgumentStack SetObjectHiliteColorOverride      (ArgumentStack&& args);
+    ArgumentStack RemoveEffectFromTURD              (ArgumentStack&& args);
+    ArgumentStack SetSpawnLocation                  (ArgumentStack&& args);
+    ArgumentStack SendDMAllCreatorLists             (ArgumentStack&& args);
+    ArgumentStack AddCustomJournalEntry             (ArgumentStack&& args);
+    ArgumentStack GetJournalEntry                   (ArgumentStack&& args);
 
     CNWSPlayer *player(ArgumentStack& args);