+++ Plugins/Tweaks/Tweaks.cpp
@@ -7,7 +7,7 @@
 #include "Tweaks/ParryAllAttacks.hpp"
 #include "Tweaks/SneakAttackCritImmunity.hpp"
 #include "Tweaks/PreserveDepletedItems.hpp"
-#include "Tweaks/HideDMsOnCharList.hpp"
+#include "Tweaks/HidePlayersOnCharList.hpp"
 #include "Tweaks/DisableMonkAbilitiesWhenPolymorphed.hpp"
 #include "Tweaks/StringToIntBaseToAuto.hpp"
 #include "Tweaks/StripOVTFromNotVisibleObject.hpp"
@@ -19,6 +19,9 @@
 #include "Tweaks/AddPrestigeclassCasterLevels.hpp"
 #include "Tweaks/FixUnlimitedPotionsBug.hpp"
 #include "Tweaks/UnhardcodeShields.hpp"
+#include "Tweaks/BlockDMSpawnItem.hpp"
+#include "Tweaks/FixArmorDexBonusUnderOne.hpp"
+#include "Tweaks/FixItemNullptrInCItemRepository.hpp"
 
 #include "Services/Config/Config.hpp"
 
@@ -89,8 +92,18 @@ Tweaks::Tweaks(Services::ProxyServiceList* services)
 
     if (GetServices()->m_config->Get<bool>("HIDE_DMS_ON_CHAR_LIST", false))
     {
-        LOG_INFO("DMs will not be visible on character list");
-        m_HideDMsOnCharList = std::make_unique<HideDMsOnCharList>(GetServices()->m_hooks.get());
+        LOG_INFO("NWNX_TWEAKS_HIDE_DMS_ON_CHAR_LIST has been deprecated, please use NWNX_TWEAKS_HIDE_PLAYERS_ON_CHAR_LIST = 1");
+        m_HidePlayersOnCharList = std::make_unique<HidePlayersOnCharList>(GetServices()->m_hooks.get(), 1);
+    }
+    else if (auto mode = GetServices()->m_config->Get<int>("HIDE_PLAYERS_ON_CHAR_LIST", 0))
+    {
+        if (mode == 1)
+            LOG_INFO("DMs will not be visible on character list.");
+        else if (mode == 2)
+            LOG_INFO("PCs will not be visible on character list.");
+        else if (mode == 3)
+            LOG_INFO("DMs and PCs will not be visible on character list.");
+        m_HidePlayersOnCharList = std::make_unique<HidePlayersOnCharList>(GetServices()->m_hooks.get(), mode);
     }
 
     if (GetServices()->m_config->Get<bool>("DISABLE_MONK_ABILITIES_WHEN_POLYMORPHED", false))
@@ -157,6 +170,24 @@ Tweaks::Tweaks(Services::ProxyServiceList* services)
     {
         LOG_INFO("Using baseitems.2da to define shield AC and create shield-like items");
         m_UnhardcodeShields = std::make_unique<UnhardcodeShields>(GetServices()->m_hooks.get());
+    }
+
+    if (GetServices()->m_config->Get<bool>("BLOCK_DM_SPAWNITEM", false))
+    {
+        LOG_INFO("Blocking the dm_spawnitem console command");
+        m_BlockDMSpawnItem = std::make_unique<BlockDMSpawnItem>(GetServices()->m_hooks.get());
+    }
+
+    if (GetServices()->m_config->Get<bool>("FIX_ARMOR_DEX_BONUS_UNDER_ONE", false))
+    {
+        LOG_INFO("Allowing armors with max DEX bonus under 1.");
+        m_FixArmorDexBonusUnderOne = std::make_unique<FixArmorDexBonusUnderOne>(GetServices()->m_hooks.get());
+    }
+
+    if (GetServices()->m_config->Get<bool>("FIX_ITEM_NULLPTR_IN_CITEMREPOSITORY", false))
+    {
+        LOG_INFO("Will check for invalid items in the CItemRepository List.");
+        m_FixItemNullptrInCItemRepository = std::make_unique<FixItemNullptrInCItemRepository>(GetServices()->m_hooks.get());
     }
 }